Create getDisplayItems helper function
